requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Telegram
python-telegram-bot==21.0.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# AI/ML
openai==1.12.0
anthropic==0.18.1
tiktoken==0.5.2

# Scheduling
apscheduler==3.10.4
redis==5.0.1

# Utilities
httpx==0.26.0
python-multipart==0.0.6
